Actually implement framebuffer driver
