+++ bbb/meta-bbb/recipes-core/trebletrouble/files/scripts/clean.sh
@@ -4,3 +4,7 @@ if [ -d bin ]; then
     rm bin/*
     rmdir bin
 fi
+
+if [ -f src/ugfx/.build/libugfx.a ]; then
+    rm src/ugfx/.build/libugfx.a
+fi
