+++ bbb/meta-bbb/recipes-core/trebletrouble/files/src/display.h
@@ -31,7 +31,4 @@ void compare_notes(int expected, int actual, int i, char* fbp);
 void clear_notes(int i, int expected[NUM_NOTES], int actual[NUM_NOTES], char* fbp);
 void load_song(FILE *song, char *fbp, int expected[NUM_NOTES]);
 
-char *init_display(int *fbfd);
-void cleanup_display(char* fbp, int *fbfd);
-
 #endif
