+++ bbb/meta-bbb/recipes-core/trebletrouble/files/src/ugfx/resources/ugfx/drivers/gdisp/framebuffer/driver.mk
@@ -1,2 +1,3 @@
 GFXINC +=	$(GFXLIB)/drivers/gdisp/framebuffer
 GFXSRC +=	$(GFXLIB)/drivers/gdisp/framebuffer/gdisp_lld_framebuffer.c
+GFXSRC +=	$(GFXLIB)/../../../gdisplay.c
